+++ api/src/models/types.ts
@@ -23,13 +23,6 @@ export type Coordinates = {
   lng: Float
 };
 
-export type Duration = Partial<{
-  seconds: number
-  minutes: number
-  hours: number
-  days: number
-}>;
-
 export type CommonTimestamps = {
   createdAt: string
   modifiedAt?: string
