+++ lib/twine-util/index.ts
@@ -1,11 +1,13 @@
+import * as Duration from './duration';
 import * as Promises from './promises';
 import * as Objects from './objects';
 import * as Random from './random';
 import * as Time from './time';
 import * as File from './file';
-import * as Url from './url'
+import * as Url from './url';
 
 export {
+  Duration,
   Promises,
   Objects,
   Random,
